refactor: centralize calls to createEvent
