+++ src/utils/event/dom-events.d.ts
@@ -1,0 +1,9 @@
+declare module '@testing-library/dom/dist/event-map.js' {
+  export const eventMap: Record<
+    string,
+    {
+      EventType: string
+      defaultInit: EventInit
+    }
+  >
+}
