+++ src/utils/event/types.ts
@@ -1,0 +1,12 @@
+export interface PointerCoords {
+  x?: number
+  y?: number
+  clientX?: number
+  clientY?: number
+  offsetX?: number
+  offsetY?: number
+  pageX?: number
+  pageY?: number
+  screenX?: number
+  screenY?: number
+}
