--- src/utils/pointer/dom-events.d.ts
@@ -1,3 +1,0 @@
-declare module '@testing-library/dom/dist/event-map.js' {
-  export const eventMap: Record<string, unknown>
-}
